`change_cwd` is not a pytest helper function
